requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
sqlalchemy==2.0.35
pydantic==2.9.2
python-dotenv==1.0.1
groq==0.11.0  # Groq SDK
pytest==8.3.3  # For tests 
